Only showing BrowserWindows when they're ready
